Refactor CosmosDb integration
This commit refactors the CosmosDb integration
as an IHttpSpanTracer implementation
