+++ src/Elastic.Apm.NetCoreAll/Elastic.Apm.NetCoreAll.csproj
@@ -16,6 +16,7 @@
     <ProjectReference Include="..\Elastic.Apm.SqlClient\Elastic.Apm.SqlClient.csproj" />
     <ProjectReference Include="..\Elastic.Apm.Azure.ServiceBus\Elastic.Apm.Azure.ServiceBus.csproj" />
     <ProjectReference Include="..\Elastic.Apm.Azure.Storage\Elastic.Apm.Azure.Storage.csproj" />
+    <ProjectReference Include="..\Elastic.Apm.Azure.CosmosDb\Elastic.Apm.Azure.CosmosDb.csproj" />
     <ProjectReference Include="..\Elastic.Apm\Elastic.Apm.csproj" />
   </ItemGroup>
 </Project>
