+++ test/Elastic.Apm.Azure.CosmosDb.Tests/Elastic.Apm.Azure.CosmosDb.Tests.csproj
@@ -23,6 +23,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\..\src\Elastic.Apm.Azure.CosmosDb\Elastic.Apm.Azure.CosmosDb.csproj" />
       <ProjectReference Include="..\..\src\Elastic.Apm\Elastic.Apm.csproj" />
       <ProjectReference Include="..\Elastic.Apm.Tests.Utilities\Elastic.Apm.Tests.Utilities.csproj" />
     </ItemGroup>
